drivers/net/dm90x0.c: Fix a syslog format

+++ drivers/net/dm90x0.c
@@ -1526,8 +1526,10 @@ static int dm9x_ifup(FAR struct net_driver_s *dev)
   int i;
 
   ninfo("Bringing up: %d.%d.%d.%d\n",
-        dev->d_ipaddr & 0xff, (dev->d_ipaddr >> 8) & 0xff,
-        (dev->d_ipaddr >> 16) & 0xff, dev->d_ipaddr >> 24);
+        (int)(dev->d_ipaddr & 0xff),
+        (int)((dev->d_ipaddr >> 8) & 0xff),
+        (int)((dev->d_ipaddr >> 16) & 0xff),
+        (int)(dev->d_ipaddr >> 24));
 
   /* Initialize DM90x0 chip */
